+++ tests/unit/test_offsets.py
@@ -1,11 +1,7 @@
-from datetime import datetime
-from types import GeneratorType
-
-import numpy as np
 import pytest
-from pandas.tseries.offsets import DateOffset, NaT
+from cftime import datetime
 
-from scmdata.offsets import apply_dt, generate_range, to_offset
+from scmdata.offsets import generate_range, to_offset
 
 
 @pytest.mark.parametrize(
@@ -35,15 +31,8 @@ def test_invalid_offsets(offset_rule):
 
 def test_annual_start():
     offset = to_offset("AS")
-    assert offset.__class__.__name__ == "LongYearBegin"
 
     dt = datetime(2001, 2, 12)
-
-    res = offset.apply(dt)
-    assert isinstance(res, datetime)
-    assert res.year == 2002
-    assert res.month == 1
-    assert res.day == 1
 
     res = offset.rollback(dt)
     assert res.year == 2001
@@ -58,15 +47,8 @@ def test_annual_start():
 
 def test_month_start():
     offset = to_offset("MS")
-    assert offset.__class__.__name__ == "LongMonthBegin"
 
     dt = datetime(2001, 2, 12)
-
-    res = offset.apply(dt)
-    assert isinstance(res, datetime)
-    assert res.year == 2001
-    assert res.month == 3
-    assert res.day == 1
 
     res = offset.rollback(dt)
     assert res.year == 2001
@@ -85,6 +67,7 @@ def test_month_start():
         [datetime(2000, 2, 12), datetime(2001, 2, 12)],
         [datetime(2000, 2, 12), datetime(3001, 2, 12)],
         [datetime(1000, 2, 12), datetime(2001, 2, 12)],
+        [datetime(2000, 2, 12), datetime(2001, 2, 12)],
     ),
 )
 def test_generate_range(start, end):
@@ -93,25 +76,17 @@ def test_generate_range(start, end):
     end = end
 
     res = generate_range(start, end, offset)
-    assert isinstance(res, GeneratorType)
-
-    dts = list(res)
-
     exp = [datetime(y, 1, 1) for y in range(start.year, end.year + 2)]
 
-    assert dts == exp
+    assert list(res) == exp
 
 
-@pytest.mark.parametrize("inp", [np.nan, NaT])
-def test_nan_apply_dt_errors(inp):
-    offset = DateOffset()
-    test_func = apply_dt(lambda _, x: x)
-    assert test_func(offset, inp) is NaT
+def test_generate_range_on_edges():
+    offset = to_offset("AS")
+    start = datetime(2000, 1, 1)
+    end = datetime(2100, 1, 1)
 
+    res = generate_range(start, end, offset)
+    exp = [datetime(y, 1, 1) for y in range(start.year, end.year + 1)]
 
-def test_nan_apply_dt_normalize():
-    offset = DateOffset(normalize=True)
-    test_func = apply_dt(lambda _, x: x)
-    assert test_func(offset, datetime(2000, 1, 1, 13, 10, 1)) == datetime(
-        2000, 1, 1, 0, 0, 0
-    )
+    assert list(res) == exp